+++ CHANGELOG.md
@@ -7,4 +7,5 @@
 Add:
 - Buckup.py
 - 404page
+- use `contentDom.val()` to get the val of the `Editor`